+++ grad_new.c
@@ -6,13 +6,13 @@
 /*   By: mdos-san <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/28 11:42:53 by mdos-san          #+#    #+#             */
-/*   Updated: 2015/12/28 11:46:07 by mdos-san         ###   ########.fr       */
+/*   Updated: 2015/12/28 14:18:04 by mdos-san         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fdf.h"
 
-t_grad	*grad_new(int r, int g, int b)
+t_grad	*grad_new(double r, double g, double b)
 {
 	t_grad	*new;
